File Details: Moved file properties tile to a separate widget
